Changed QObject relative inheritance path
It was left up to concrete class implementations of ITalkAsyncToListener to
inherit QObject.

However, the interface, and its parent interface, IRunConcurrent both make use
of signal and slot members. Because of this, they need to be decendants of
QObject in order for any connect statement that uses them to be valid.

I am still calling the QOBJECT macro in the concrete class. We will attempt an
implementation to see if it links correctly.